Lexing is basically done. To add new tokens we just need to add the relevant entries in lexer.mll
For now we're using ocamllex in order to get to get to something resembling a
complete compiler faster. Once we're at that point, we'll come back and implement
our own lexer and add improvements and extensions to other parts of the compiler.
Making sure we respect modularity will make that reasonable to do