Fixed funcion decl by requiring functions start with special character
No idea how to fix these ambiguities. I have a full page of purple
links from google that mostly seem to say that we are doing this
correctly. SE also seems to fail in giving me a good reason why this is
wrong.
In particular the current error is that a*b*c is ambiguous, but the
TIMES token is stated t obe left associative in our .mly file. How is
that ambiguous? Need to work on 179 for tm.