Added tests for detecting scope and type errors
Have yet to add the test to check against false errors, but that's the
easy one.

Fixed line numbers so that contiguous newlines are counted
individually.
Fixed an error where ok function calls were though to have the incorrect
number of arguments, and fixed an error where the type checker would
only think two arrays or two functions had the same type if the two were
literally the same object as each other.