Added return value type checking
Ended up changing how functions are declared so that the type must be
given in the declaration (return values are still checked to match
against it). That was a mistake since it only added to the logic, but
at least it makes programs in the language clearer.

Also added the intermediate code tree definition